requirements.txt
app==0.0.1
asgiref==3.5.0
Authlib==1.0.1
beautifulsoup4==4.11.1
certifi==2021.10.8
cffi==1.15.0
cfgv==3.3.1
charset-normalizer==2.0.12
click==8.1.2
coverage==6.3.2
cryptography==36.0.2
distlib==0.3.4
Django==4.0.4
django-extensions==3.1.5
docopt==0.6.2
factory_boy==3.2.1
filelock==3.6.0
identify==2.4.12
idna==3.3
nodeenv==1.6.0
nose2==0.11.0
oauthlib==3.2.0
platformdirs==2.5.1
pre-commit==2.18.1
pycparser==2.21
python-dotenv==0.20.0
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
soupsieve==2.3.2
sqlparse==0.4.2
toml==0.10.2
urllib3==1.26.9
virtualenv==20.14.1
yarg==0.1.9
